+++ src/main/java/com/gmail/netcracker/application/dto/dao/interfaces/NoteDao.java
@@ -11,7 +11,7 @@ public interface NoteDao {
 
     Note getNote(int noteId);
 
-    void delete(int noteId);
+    void delete(Long noteId);
 
     void update(Note note);
