+++ src/main/java/com/gmail/netcracker/application/service/interfaces/NoteService.java
@@ -10,7 +10,7 @@ public interface NoteService {
 
     Note getNote(int noteId);
 
-    void delete(int noteId);
+    void delete(Long noteId);
 
     void update(Note note);
